requirements.txt
python-dotenv
requests
httpx
openai==1.11.0
python-telegram-bot==20.7
